+++ apps/web/src/components/forms/LoginForm.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import Button from "./ui/Button";
-import Input from "./ui/Input";
-import { api } from "../utils/api";
-import { Form, FormActions } from "./ui/Form";
+import Button from "../ui/Button";
+import Input from "../ui/Input";
+import { api } from "../../utils/api";
+import { Form, FormActions } from "../ui/Form";
 
 export default function LoginForm() {
   const utils = api.useUtils();
