+++ apps/web/src/routes/Register.tsx
@@ -1,5 +1,5 @@
 import CenteredPage from "../components/CenteredPage";
-import RegisterForm from "../components/RegisterForm";
+import RegisterForm from "../components/forms/RegisterForm";
 import Card from "../components/ui/Card";
 
 export default function Register() {
